Explore the results of the integration analysis between miRNAs and genes

After performing the integration analysis between miRNA and gene expression values with the mirnaIntegration() function, the results are stored in the integration slot of a MirnaExperiment object and can be explored with this function.

integration(object, param = FALSE)

Arguments

object

A MirnaExperiment object containing miRNA and gene data

param

Logical, whether to return the complete list object with the parameters used, or just the results stored in data. Default is FALSE

Value

If param is FALSE, then this functions returns a data.frame object containing the results of the integration analysis. Otherwise, a list

object including the parameters used for the analysis will be returned.

Author

Jacopo Ronchi, jacopo.ronchi@unimib.it

Examples

# load example MirnaExperiment object
obj <- loadExamples()

# perform Kendall's correlation analysis with tau > 0.8 and p < 0.05
obj <- mirnaIntegration(obj,
    test = "correlation",
    corMethod = "kendall", corCutoff = 0.8
)
#> As specified by the user, a correlation will be used.
#> Performing Kendall's correlation analysis...
#> A statistically significant correlation between 1 miRNA-target pairs was found!

# visualize the results of correlation analysis
res <- integration(obj)
res
#>                      microRNA Target microRNA.Direction Corr.Coefficient
#> hsa.miR.21.5p.8 hsa-miR-21-5p  MATN2        upregulated       -0.8166667
#>                 Corr.P.Value Corr.Adjusted.P.Val
#> hsa.miR.21.5p.8 5.116119e-06         0.003289664