Extract the database used for integrative miRNA-mRNA pathway analyses

This function accesses the database slot of a FunctionalEnrichment object and returns the name of the database used by the topologicalAnalysis() function to perform the integrative topological analysis.

Usage

integrationDatabase(object)

Arguments

object

An object of class IntegrativePathwayAnalysis containing the results of a miRNA-mRNA pathway analysis

Value

A character object with the name of the database used by the topologicalAnalysis() function, such as KEGG.

Author

Jacopo Ronchi, jacopo.ronchi@unimib.it

Examples

# load the example IntegrativePathwayAnalysis object
obj <- loadExamples("IntegrativePathwayAnalysis")

# see the database
integrationDatabase(obj)
#> [1] "KEGG"