List all the available biological pathways in KEGG, Reactome and WikiPathways

This function can be used to retrieve a list of valid biological pathways present in KEGG, Reactome and WikiPathways.

listPathways(organism, database)

Arguments

organism

The name of the organism under consideration. The different databases have different supported organisms. To see the list of supported organisms for a given database, use the supportedOrganisms() function

database

The name of the database to use. It must be one of: KEGG, Reactome, and WikiPathways

Value

A character vector containing the pathway names present in the specified database.

Note

This function uses the graphite package to retrieve biological pathways from KEGG, Reactome and WikiPathways.

References

Sales, G., Calura, E., Cavalieri, D. et al. graphite - a Bioconductor package to convert pathway topology to gene network. BMC Bioinformatics 13, 20 (2012), https://doi.org/10.1186/1471-2105-13-20.

Author

Jacopo Ronchi, jacopo.ronchi@unimib.it

Examples

# list the mouse pathways present in WikiPathways
listPathways("Mus musculus", "WikiPathways")
#>   [1] "Statin pathway"                                                                  
#>   [2] "Cholesterol biosynthesis"                                                        
#>   [3] "Selenium metabolism / selenoproteins"                                            
#>   [4] "TGF-beta signaling pathway"                                                      
#>   [5] "Hedgehog signaling pathway"                                                      
#>   [6] "Glucuronidation"                                                                 
#>   [7] "EBV LMP1 signaling"                                                              
#>   [8] "Estrogen signaling"                                                              
#>   [9] "Transcriptional activation by Nfe2l2 in response to phytochemicals"              
#>  [10] "Methylation"                                                                     
#>  [11] "EPO receptor signaling"                                                          
#>  [12] "Amino acid conjugation of benzoic acid"                                          
#>  [13] "Type II interferon signaling (IFNG)"                                             
#>  [14] "Apoptosis"                                                                       
#>  [15] "Nod-like receptor (NLR) signaling pathway"                                       
#>  [16] "Retinol metabolism"                                                              
#>  [17] "ErbB signaling pathway"                                                          
#>  [18] "Aflatoxin B1 metabolism"                                                         
#>  [19] "Mitochondrial gene expression"                                                   
#>  [20] "Estrogen metabolism"                                                             
#>  [21] "Polyol pathway"                                                                  
#>  [22] "SIDS susceptibility pathways"                                                    
#>  [23] "Endochondral ossification"                                                       
#>  [24] "Selenium micronutrient network"                                                  
#>  [25] "Folic acid network"                                                              
#>  [26] "Oxidation by cytochrome P450"                                                    
#>  [27] "Oxidative damage response"                                                       
#>  [28] "Dopaminergic neurogenesis"                                                       
#>  [29] "Regulation of cardiac hypertrophy by miR-208"                                    
#>  [30] "MicroRNAs in cardiomyocyte hypertrophy"                                          
#>  [31] "Glycolysis and gluconeogenesis"                                                  
#>  [32] "Iron homeostasis"                                                                
#>  [33] "Cytoplasmic ribosomal proteins"                                                  
#>  [34] "Glutathione metabolism"                                                          
#>  [35] "Apoptosis modulation by HSP70"                                                   
#>  [36] "Acetylcholine synthesis"                                                         
#>  [37] "Mechanisms associated with pluripotency"                                         
#>  [38] "One-carbon metabolism and related pathways"                                      
#>  [39] "Kennedy pathway"                                                                 
#>  [40] "Heme biosynthesis"                                                               
#>  [41] "GPCRs, class A rhodopsin-like"                                                   
#>  [42] "Splicing factor NOVA regulated synaptic proteins"                                
#>  [43] "Complement activation, classical pathway"                                        
#>  [44] "Ptf1a related regulatory pathway"                                                
#>  [45] "Hypertrophy model"                                                               
#>  [46] "Heart development"                                                               
#>  [47] "Neural crest differentiation"                                                    
#>  [48] "Alzheimer's disease"                                                             
#>  [49] "Serotonin receptor 2 and STAT3 signaling"                                        
#>  [50] "SREBF and miR33 in cholesterol and lipid homeostasis"                            
#>  [51] "Serotonin and anxiety-related events"                                            
#>  [52] "Serotonin and anxiety"                                                           
#>  [53] "BDNF pathway"                                                                    
#>  [54] "Purine metabolism"                                                               
#>  [55] "Chemokine signaling pathway"                                                     
#>  [56] "PPAR signaling pathway"                                                          
#>  [57] "Fatty acid oxidation"                                                            
#>  [58] "G protein signaling pathways"                                                    
#>  [59] "miRNAs and TFs in iPS Cell Generation"                                           
#>  [60] "Osteoblast signaling"                                                            
#>  [61] "Spinal cord injury"                                                              
#>  [62] "Mapk cascade"                                                                    
#>  [63] "Primary focal segmental glomerulosclerosis (FSGS)"                               
#>  [64] "Focal adhesion: PI3K-Akt-mTOR signaling pathway"                                 
#>  [65] "Gene regulatory network modelling somitogenesis"                                 
#>  [66] "White fat cell differentiation"                                                  
#>  [67] "Notch signaling pathway"                                                         
#>  [68] "Electron transport chain"                                                        
#>  [69] "G13 signaling pathway"                                                           
#>  [70] "Translation factors"                                                             
#>  [71] "Glycogen metabolism"                                                             
#>  [72] "Eicosanoid synthesis"                                                            
#>  [73] "Fatty acid omega-oxidation"                                                      
#>  [74] "ESC pluripotency pathways"                                                       
#>  [75] "p38 Mapk signaling pathway"                                                      
#>  [76] "ApoE and miR-146 in inflammation and atherosclerosis"                            
#>  [77] "Tyrobp causal network in microglia"                                              
#>  [78] "Microglia pathogen phagocytosis pathway"                                         
#>  [79] "Lung fibrosis"                                                                   
#>  [80] "Parkinson's disease"                                                             
#>  [81] "EDA signaling in hair follicle development"                                      
#>  [82] "Novel Jun-Dmp1 pathway"                                                          
#>  [83] "BMP signaling pathway in eyelid development"                                     
#>  [84] "Hfe effect on hepcidin production"                                               
#>  [85] "Factors and pathways affecting insulin-like growth factor (IGF1)-Akt signaling"  
#>  [86] "IL-1 signaling pathway"                                                          
#>  [87] "Prostaglandin synthesis and regulation"                                          
#>  [88] "Myometrial relaxation and contraction pathways"                                  
#>  [89] "Wnt signaling in kidney disease"                                                 
#>  [90] "Robo4 and VEGF signaling pathways crosstalk"                                     
#>  [91] "ACE inhibitor pathway"                                                           
#>  [92] "miR-127 in mesendoderm differentiation"                                          
#>  [93] "Wnt signaling pathway"                                                           
#>  [94] "Oxidative stress response"                                                       
#>  [95] "G1 to S cell cycle control"                                                      
#>  [96] "Distal convoluted tubule 1 (DCT1) cell"                                          
#>  [97] "Ethanol metabolism resulting in production of ROS by CYP2E1"                     
#>  [98] "Nuclear receptors in lipid metabolism and toxicity"                              
#>  [99] "Eicosanoid lipid synthesis map"                                                  
#> [100] "TCA cycle"                                                                       
#> [101] "Sphingolipid metabolism overview"                                                
#> [102] "Glycerolipids and glycerophospholipids"                                          
#> [103] "Cholesterol metabolism with Bloch and Kandutsch-Russell pathways"                
#> [104] "Eicosanoid metabolism via cyclooxygenases (COX)"                                 
#> [105] "Eicosanoid metabolism via lipoxygenases (LOX)"                                   
#> [106] "Eicosanoid metabolism via cytochrome P450 monooxygenases"                        
#> [107] "One-carbon metabolism"                                                           
#> [108] "Omega-3 / omega-6 fatty acid synthesis"                                          
#> [109] "Omega-9 fatty acid synthesis"                                                    
#> [110] "Oxidative stress and redox pathway"                                              
#> [111] "Circulating monocytes and cardiac macrophages in diastolic dysfunction"          
#> [112] "Osteoclast signaling"                                                            
#> [113] "Inflammatory response pathway"                                                   
#> [114] "Blood clotting cascade"                                                          
#> [115] "Lipids measured in liver metastasis from breast cancer"                          
#> [116] "Sphingolipid metabolism (integrated pathway)"                                    
#> [117] "Regulation of Pgc1a expression by a Gsk3b-Tfeb signaling axis in skeletal muscle"
#> [118] "GDNF/RET signaling axis"                                                         
#> [119] "Peroxiredoxin 2 induced ovarian failure"                                         
#> [120] "Mapk signaling pathway"                                                          
#> [121] "Deregulation of renin-angiotensin system by SARS-CoV infection"                  
#> [122] "Hypoxia-dependent self-renewal of myoblasts"                                     
#> [123] "Hypoxia-dependent proliferation of myoblasts"                                    
#> [124] "Hypoxia-dependent differentiation of myoblasts"                                  
#> [125] "Burn wound healing"                                                              
#> [126] "Fibrin complement receptor 3 signaling pathway"                                  
#> [127] "Proteasome degradation"                                                          
#> [128] "Biogenic amine synthesis"                                                        
#> [129] "Regulation of actin cytoskeleton"                                                
#> [130] "Synthesis and degradation of ketone bodies"                                      
#> [131] "Exercise-induced circadian regulation"                                           
#> [132] "Steroid biosynthesis"                                                            
#> [133] "Calcium regulation in cardiac cells"                                             
#> [134] "Signal transduction of S1P receptor"                                             
#> [135] "FAS pathway and stress induction of HSP regulation"                              
#> [136] "Leptin-insulin signaling overlap"                                                
#> [137] "Integrin-mediated cell adhesion"                                                 
#> [138] "miR-1 in cardiac development"                                                    
#> [139] "Pentose phosphate pathway"                                                       
#> [140] "Insulin signaling"                                                               
#> [141] "Amino acid metabolism"                                                           
#> [142] "Leptin and adiponectin"                                                          
#> [143] "Wnt signaling pathway and pluripotency"                                          
#> [144] "Glutathione and one-carbon metabolism"                                           
#> [145] "Focal adhesion"                                                                  
#> [146] "Toll-like receptor signaling"                                                    
#> [147] "Oxidative phosphorylation"                                                       
#> [148] "Arachidonate epoxygenase / epoxide hydrolase"                                    
#> [149] "Metapathway biotransformation"                                                   
#> [150] "Fatty acid beta-oxidation"                                                       
#> [151] "Fatty acid biosynthesis"                                                         
#> [152] "Tryptophan metabolism"                                                           
#> [153] "Nucleotide GPCRs"                                                                
#> [154] "GPCRs, small ligand"                                                             
#> [155] "Monoamine GPCRs"